refactor(api): standardise to kebab-case naming
