+++ apps/api/src/features/cars/cars.service.ts
