+++ apps/api/src/lib/__tests__/get-latest-month.test.ts
@@ -1,6 +1,6 @@
 import { db } from "@sgcarstrends/database";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { getLatestMonth } from "../getLatestMonth";
+import { getLatestMonth } from "../get-latest-month";
 
 vi.mock("@sgcarstrends/database", () => ({
   db: {
