+++ apps/api/src/lib/__tests__/get-months-by-year.test.ts
@@ -1,12 +1,12 @@
 import { describe, expect, it } from "vitest";
-import { groupMonthsByYear } from "../groupMonthsByYear";
+import { getMonthsByYear } from "../get-months-by-year";
 
-describe("groupMonthsByYear", () => {
+describe("getMonthsByYear", () => {
   it("should group months by year", () => {
     const months = ["2022-01", "2022-12", "2023-01", "2023-02"];
 
-    expect(Object.keys(groupMonthsByYear(months))).toHaveLength(2);
-    expect(groupMonthsByYear(months)).toEqual([
+    expect(Object.keys(getMonthsByYear(months))).toHaveLength(2);
+    expect(getMonthsByYear(months)).toEqual([
       { year: "2023", months: ["02", "01"] },
       { year: "2022", months: ["12", "01"] },
     ]);
@@ -15,7 +15,7 @@ describe("groupMonthsByYear", () => {
   it("should handle empty array", () => {
     const months: string[] = [];
 
-    const result = groupMonthsByYear(months);
+    const result = getMonthsByYear(months);
 
     expect(Object.keys(result)).toHaveLength(0);
   });
@@ -23,8 +23,8 @@ describe("groupMonthsByYear", () => {
   it("should handle single year", () => {
     const months = ["2023-01", "2023-02", "2023-03"];
 
-    expect(Object.keys(groupMonthsByYear(months))).toHaveLength(1);
-    expect(groupMonthsByYear(months)).toEqual([
+    expect(Object.keys(getMonthsByYear(months))).toHaveLength(1);
+    expect(getMonthsByYear(months)).toEqual([
       {
         year: "2023",
         months: ["03", "02", "01"],
