+++ apps/api/src/lib/get-unique-months.ts
